Add files via upload
The first step is to read the image in RGB format and then convert it into a grayscale image. This will turn an image into a classic B&W image. The next step is to invert the B&W image, aka negative image. Inversion can be used to enhance details. Then we can finally create the pencil sketch by mixing the grayscale image with the inverted blurry image. This can be done by dividing the grayscale image by the inverted blurry image. Since images are just arrays, we can easily do this by using the divide function from the CV2 library in Python.